requirements.txt
apache-airflow==2.7.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.1
lxml==4.9.3
fake-useragent==1.4.0
python-dateutil==2.8.2 